@@ -1,19 +1,18 @@
 package pl.piomin.services.protobuf.customer;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.feign.EnableFeignClients;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.converter.protobuf.ProtobufHttpMessageConverter;
 import org.springframework.web.client.RestTemplate;
-
 import pl.piomin.services.protobuf.customer.data.CustomerRepository;
 import pl.piomin.services.protobuf.customer.model.CustomerProto.Customer;
 import pl.piomin.services.protobuf.customer.model.CustomerProto.Customer.CustomerType;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 @SpringBootApplication
 @EnableFeignClients
